[keystone] Option for central hermes rabbitmq
This provides the option of using the HA setup rabbitmq from hermes
instead of the single non-ha one.

To ensure that no messages are dropped, one first needs to
supply audit.user and audit.password to enable
sending out the new events and after ensuring that
no messages are on the queue of the old rabbitmq,
one can disable the internal one.